% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{attenuate_xb}
\alias{attenuate_xb}
\title{Return marginal expectation with attenuation}
\usage{
attenuate_xb(xb, Z, D, link)
}
\arguments{
\item{xb}{Vector of values of the linear predictor}

\item{Z}{Random effects design matrix}

\item{D}{Covariance matrix of the random effects}

\item{link}{String specifying the link function}
}
\value{
A vector
}
\description{
The marginal expectation may be better approximated using an attenuation 
scheme in non-linear models. This function returns the attenuated linear predictor.
Used internally.
}
