% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Betadist.R
\name{Beta}
\alias{Beta}
\title{Beta distribution declaration}
\usage{
Beta(link = "logit")
}
\arguments{
\item{link}{Name of link function. Only accepts `logit` currently.}
}
\value{
A list with two elements naming the family and link function
}
\description{
Skeleton list to declare a Beta distribution in a `Model` object
}
