% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{blockMat}
\alias{blockMat}
\title{Combines a field of matrices into a block diagonal matrix}
\usage{
blockMat(matfield)
}
\arguments{
\item{matfield}{A field of matrices}
}
\value{
A block diagonal matrix
}
\description{
Combines a field of matrices into a block diagonal matrix. Used on
the output of `genD`
}
