% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfunctions.R
\name{coef.mcml}
\alias{coef.mcml}
\title{Extracts coefficients from a mcml object}
\usage{
\method{coef}{mcml}(object, ...)
}
\arguments{
\item{object}{An `mcml` model fit.}

\item{...}{Further arguments passed from other methods}
}
\value{
A data frame summarising the parameters including the random effects.
}
\description{
Extracts the coefficients from an `mcml` object returned from a call of `MCML` or `LA` in the \link[glmmrBase]{Model} class.
}
