% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_s3_functions.R
\name{coef.Model}
\alias{coef.Model}
\title{Extracts coefficients from a Model object}
\usage{
\method{coef}{Model}(object, ...)
}
\arguments{
\item{object}{A `Model` object.}

\item{...}{Further arguments passed from other methods}
}
\value{
Fixed effect and covariance parameters extracted from the model object.
}
\description{
Extracts the coefficients from a `Model` object.
}
