% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfunctions.R
\name{confint.mcml}
\alias{confint.mcml}
\title{Fixed effect confidence intervals for a `mcml` object}
\usage{
\method{confint}{mcml}(object, ...)
}
\arguments{
\item{object}{A `mcml` object.}

\item{...}{Further arguments passed from other methods}
}
\value{
A matrix (or vector) with columns giving lower and upper confidence limits for each parameter.
}
\description{
Returns the computed confidence intervals for a `mcml` object.
}
