% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelder.R
\name{cross_df}
\alias{cross_df}
\title{Generate crossed block structure}
\usage{
cross_df(df1, df2)
}
\arguments{
\item{df1}{data frame}

\item{df2}{data frame}
}
\value{
data frame
}
\description{
Generate a data frame with crossed rows from two other data frames
}
\details{
For two data frames `df1` and `df2`, the function will return another data frame
that crosses them, which has rows with every unique combination of the input data frames
}
\examples{
cross_df(data.frame(t=1:4),data.frame(cl=1:3))
}
