% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelder.R
\name{cycles}
\alias{cycles}
\title{Generates all the orderings of a}
\usage{
cycles(a)
}
\arguments{
\item{a}{vector}
}
\value{
vector
}
\description{
Given input a, returns a length(a)^2 vector by cycling through the values of a
}
