% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_funcs.R
\name{dfexp}
\alias{dfexp}
\title{Exponential function}
\usage{
dfexp(x)
}
\arguments{
\item{x}{List with named elements `data`, a vector of covariate values, and `pars`, a vector of two parameters}
}
\description{
Exponential function
}
