% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlinkdeta}
\alias{dlinkdeta}
\title{Partial derivative of link function with respect to linear predictor}
\usage{
dlinkdeta(xb, link)
}
\arguments{
\item{xb}{Vector of values of the linear predictor}

\item{link}{String specifying the link function}
}
\value{
A vector
}
\description{
Returns the partial derivative of link function with respect to linear predictor.
Used internally.
}
