% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Betadist.R
\name{exponential}
\alias{exponential}
\title{Exponential distribution declaration}
\usage{
exponential(link = "logit")
}
\arguments{
\item{link}{Name of link function. Only accepts `logit` currently.}
}
\value{
A list with two elements naming the family and link function
}
\description{
Skeleton list to declare an exponential distribution in a `Model` object
}
