% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfunctions.R
\name{extractAIC.mcml}
\alias{extractAIC.mcml}
\title{Extracts the computed AIC from an mcml object}
\usage{
\method{extractAIC}{mcml}(object, ...)
}
\arguments{
\item{object}{An `mcml` model fit.}

\item{...}{Further arguments passed from other methods}
}
\value{
A numeric value.
}
\description{
Extracts the conditional Akaike Information Criterion from an mcml object returned from call of `MCML` or `LA` in the \link[glmmrBase]{Model} class.
}
