% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfunctions.R
\name{family.mcml}
\alias{family.mcml}
\title{Extracts the family from a `mcml` object.}
\usage{
\method{family}{mcml}(object, ...)
}
\arguments{
\item{object}{A `mcml` object.}

\item{...}{Further arguments passed from other methods}
}
\value{
A \link[stats]{family} object.
}
\description{
Extracts the \link[stats]{family} from a `mcml` object.
}
