% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfunctions.R
\name{fitted.mcml}
\alias{fitted.mcml}
\title{Fitted values from a `mcml` object}
\usage{
\method{fitted}{mcml}(object, ...)
}
\arguments{
\item{object}{A `mcml` object.}

\item{...}{Further arguments passed from other methods}
}
\value{
Nothing, called for effects, unless `override` is TRUE, when it will return a vector of fitted values.
}
\description{
Fitted values should not be generated directly from an `mcml` object, rather fitted values should be
generated using the original `Model`. A message is printed to the user.
}
