% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_s3_functions.R
\name{fitted.Model}
\alias{fitted.Model}
\title{Extract or generate fitted values from a `Model` object}
\usage{
\method{fitted}{Model}(object, ...)
}
\arguments{
\item{object}{A `Model` object.}

\item{...}{Further arguments passed from other methods}
}
\value{
Fitted values
}
\description{
Return fitted values. Does not account for the random effects. This function is a wrapper for `Model$fitted()`, which 
also provides a variety of additional options for generating fitted values from mixed models. 
For simulated values based on resampling random effects, see also `Model$sim_data()`. To predict the values including random effects at a new location see also
`Model$predict()`.
}
