% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfunctions.R
\name{fixed.effects}
\alias{fixed.effects}
\title{Extracts the fixed effect estimates}
\usage{
fixed.effects(object, ...)
}
\arguments{
\item{object}{An `mcml` model fit.}
}
\value{
A named, numeric vector of fixed-effects estimates.
}
\description{
Extracts the fixed effect estimates from an mcml object returned from call of `MCML` or `LA` in the \link[glmmrBase]{Model} class.
}
