% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfunctions.R
\name{formula.mcml}
\alias{formula.mcml}
\title{Extracts the formula from a `mcml` object.}
\usage{
\method{formula}{mcml}(object, fixed = TRUE, ...)
}
\arguments{
\item{object}{A `mcml` object.}

\item{fixed}{Logical. If TRUE the formula for the fixed effects is returned, otherwise the formula
for the random effects is returned.}

\item{...}{Further arguments passed from other methods}
}
\value{
A \link[stats]{formula} object.
}
\description{
Extracts the \link[stats]{formula} from a `mcml` object. Separate formulae are 
specified for the fixed and random effects in the model, either of which can be
returned. The complete formula is available from the generating `Model` object as 
`Model$formula` or `formula(Model)`
}
