% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_s3_functions.R
\name{formula.Model}
\alias{formula.Model}
\title{Extracts the formula from a `Model` object}
\usage{
\method{formula}{Model}(x, ...)
}
\arguments{
\item{x}{A `Model` object.}

\item{...}{Further arguments passed from other methods}
}
\value{
A \link[stats]{formula} object.
}
\description{
Extracts the \link[stats]{formula} from a `Model` object. This information can also be
accessed directly from the Model as `Model$formula`
}
