% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_dhdmu}
\alias{gen_dhdmu}
\title{Generates the derivative of the link function with respect to the mean. Used internally in the Model function class.}
\usage{
gen_dhdmu(xb, family, link)
}
\arguments{
\item{xb}{Vector with mean function value evaluated at fitted model parameters}

\item{family}{String declaring model family}

\item{link}{String declaring model link function}
}
\value{
Vector of derivative values
}
\description{
Generates the derivative of the link function with respect to the mean. Used internally in the Model function class.
}
