% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gen_sigma_approx}
\alias{gen_sigma_approx}
\title{Generates an approximation to the covariance of y}
\usage{
gen_sigma_approx(xb, Z, D, family, link, var_par, attenuate, qlik = TRUE)
}
\arguments{
\item{xb}{Vector of values of the linear predictor}

\item{Z}{Random effects design matrix}

\item{D}{Covariance matrix of the random effects}

\item{family}{String specifying the family}

\item{link}{String specifying the link function}

\item{var_par}{Value of the optional scale parameter}

\item{attenuate}{Logical indicating whether to use "attenuated" values of the linear predictor}

\item{qlik}{Not used.}
}
\value{
A matrix
}
\description{
Generates a first-order approximation to the covariance matrix 
of y based on the marginal quasi-likelihood. This approximation is
exact for the Gaussian-identity model. Used internally by the \link[glmmrBase]{Model} class.
}
