% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{genCholD}
\alias{genCholD}
\title{Generates the cholesky decomposition covariance matrix of the random effects}
\usage{
genCholD(D_data, gamma)
}
\arguments{
\item{D_data}{Named list specifying the covariance matrix D. Usually the return from the member function `get_D_data()` of the 
covariance class}

\item{gamma}{Vector of covariance parameters specified in order they appear column wise in the functions 
specified by `func_def`}
}
\value{
A lower triangular matrix matrix
}
\description{
Generates the covariance matrix of the random effects from a sparse representation
}
