% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{genX}
\alias{genX}
\title{Generates the matrix X}
\usage{
genX(formula, data, colnames)
}
\arguments{
\item{formula}{A string specifying the formula}

\item{data}{A matrix with the data}

\item{colnames}{A vector of strings specifying the column names of the data}
}
\value{
A symmetric positive definite covariance matrix
}
\description{
Generates the matrix X from a formula and data. This function replicates the functionality of the 
R function model.matrix, but is intended to be extended to accomodate linearisation of non-linear functions
and other functionality in future versions.
}
