% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{genZ}
\alias{genZ}
\title{Generates the design matrix of the random effects}
\usage{
genZ(formula, data, colnames)
}
\arguments{
\item{formula}{A string specifying the formula}

\item{data}{A matrix with the data}

\item{colnames}{A vector of strings specifying the column names of the data}
}
\value{
A symmetric positive definite covariance matrix
}
\description{
Generates the design matrix of the random effects.
}
