\name{glmmrBase-package}
\alias{glmmrBase-package}
\alias{glmmrBase}
\docType{package}
\title{
\packageTitle{glmmrBase}
}
\description{
\packageDescription{glmmrBase}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{glmmrBase}
\packageIndices{glmmrBase}
https://github.com/samuel-watson/glmmrBase/blob/master/README.md
}
\author{
\packageAuthor{glmmrBase}

Maintainer: \packageMaintainer{glmmrBase}
}
\keyword{ package }


