% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrafunctions.R
\name{gr}
\alias{gr}
\title{Group indicator covariance function}
\usage{
gr(x)
}
\arguments{
\item{x}{A list with named elements `pars` and `data`. `pars` is a vector with one parameter value, and
`data` is the data `x`}
}
\value{
vector of values of the function
}
\description{
Group indicator covariance function
}
\details{
The function:

\dexp{f(x) = 1(x==0)*\theta_1}
}
\examples{
fexp(list(pars = c(0.8),data=c(0,1,0,2,3,4,0)))
}
