% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{n_cov_pars}
\alias{n_cov_pars}
\title{Returns the number of covariance parameters required for the formula}
\usage{
n_cov_pars(formula, data, colnames)
}
\arguments{
\item{formula}{A string specifying the formula}

\item{data}{A matrix with the data}

\item{colnames}{A vector of strings specifying the column names of the data}
}
\value{
Integer count of parameters
}
\description{
Returns the number of covariance parameters required for the formula
}
