% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nelder.R
\name{nest_df}
\alias{nest_df}
\title{Generate nested block structure}
\usage{
nest_df(df1, df2)
}
\arguments{
\item{df1}{data frame}

\item{df2}{data frame}
}
\value{
data frame
}
\description{
Generate a data frame that nests one data frame in another
}
\details{
For two data frames `df1` and `df2`, the function will return another data frame
that nests `df2` in `df1`. So each row of `df1` will be duplicated `nrow(df2)` times 
and matched with `df2`. The values of each `df2` will be unique for each row of `df1`
}
\examples{
nest_df(data.frame(t=1:4),data.frame(cl=1:3))
}
