% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfunctions.R
\name{predict.mcml}
\alias{predict.mcml}
\title{Predict from a `mcml` object}
\usage{
\method{predict}{mcml}(object, ...)
}
\arguments{
\item{object}{A `mcml` object.}

\item{...}{Further arguments passed from other methods}
}
\value{
Nothing. Called for effects.
}
\description{
Predictions cannot be generated directly from an `mcml` object, rather new predictions should be
generated using the original `Model`. A message is printed to the user.
}
