% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrafunctions.R
\name{progress_bar}
\alias{progress_bar}
\title{Generates a progress bar}
\usage{
progress_bar(i, n, len = 30)
}
\arguments{
\item{i}{integer. The current iteration.}

\item{n}{integer. The total number of interations}

\item{len}{integer. Length of the progress a number of characters}
}
\value{
A character string
}
\description{
Prints a progress bar
}
\examples{
progress_bar(10,100)
}
