% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiledist.R
\name{Quantile}
\alias{Quantile}
\title{Family declaration to support quantile regression}
\usage{
Quantile(link = "identity", scaled = FALSE, q = 0.5)
}
\arguments{
\item{link}{Name of the link function - any of `identity`, `log`, `logit`, `inverse`, or `probit`}

\item{scaled}{Logical indicating whether to include a scale parameter. If FALSE then the scale parameter is one.}

\item{q}{Scalar in [0,1] declaring the quantile of interest.}
}
\value{
A list with two elements naming the family and link function
}
\description{
Skeleton list to declare a quantile regression model in a `Model` object.
}
