% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfunctions.R
\name{random.effects}
\alias{random.effects}
\title{Extracts the random effect estimates}
\usage{
random.effects(object, ...)
}
\arguments{
\item{object}{An `mcml` model fit.}
}
\value{
A matrix of dimension (number of fixed effects ) x (number of MCMC samples). For Laplace approximation, the number of "samples" equals one.
}
\description{
Extracts the random effect estimates or samples from an mcml object returned from call of `MCML` or `LA` in the \link[glmmrBase]{Model} class.
}
