% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfunctions.R
\name{ranef.mcml}
\alias{ranef.mcml}
\title{Extracts the random effect estimates}
\usage{
\method{ranef}{mcml}(object, ...)
}
\arguments{
\item{object}{An `mcml` model fit.}

\item{...}{Further arguments passed from other methods}
}
\value{
A matrix of dimension (number of fixed effects ) x (number of MCMC samples). For Laplace approximation, the number of "samples" equals one.
}
\description{
Extracts the random effect estimates or samples from an mcml object returned from call of `MCML` or `LA` in the \link[glmmrBase]{Model} class.
}
