% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{re_names}
\alias{re_names}
\title{Gets the names of the random effects terms}
\usage{
re_names(formula)
}
\arguments{
\item{formula}{A string specifying the formula}
}
\value{
A vector of variable names
}
\description{
Gets the names of the random effects terms
}
