% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrafunctions.R
\name{relist}
\alias{relist}
\title{Generate a list of a given structure with new values}
\usage{
relist(lst, value, p = 0)
}
\arguments{
\item{lst}{List whose structure will be replicated with the new values}

\item{value}{vector of values to replace in `lst`. The order of replacement is given by `unlist(lst)`}

\item{p}{Used internally to track level of recursion}
}
\value{
A list of the same structure as `lst` but with the values given by `value`
}
\description{
Generate a list of a given structure with new values
}
\examples{
df <- nelder(~(cl(5)*t(5)) > ind(5))
cov <- Covariance$new(formula = ~(1|gr(j)*pexp(t)),
                      parameters = list(list(0.05,0.8)),
                      data= df)
cov$parameters <- relist(cov$parameters,
                         c(0.01,0.5))
}
