% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrafunctions.R
\docType{data}
\name{Salamanders}
\alias{Salamanders}
\title{Salamanders data}
\description{
Obtained from 
\code{
uu <- url("http://www.math.mcmaster.ca/bolker/R/misc/salamander.txt")
sdat <- read.table(uu,header=TRUE,colClasses=c(rep("factor",5),"numeric"))
}
See \url{https://rpubs.com/bbolker/salamander} for more information.
}
