% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setParallel}
\alias{setParallel}
\title{Disable or enable parallelised computing}
\usage{
setParallel(parallel_, cores_ = 2L)
}
\arguments{
\item{parallel_}{Logical indicating whether to use parallel computation (TRUE) or disable it (FALSE)}

\item{cores_}{Number of cores for parallel execution}
}
\value{
None, called for effects
}
\description{
By default, the package will use multithreading for many calculations if OpenMP is 
available on the system. For multi-user systems this may not be desired, so parallel
execution can be disabled with this function.
}
