% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_s3_functions.R
\name{summary.Model}
\alias{summary.Model}
\title{Summarizes a `Model` object}
\usage{
\method{summary}{Model}(object, max_n = 10, ...)
}
\arguments{
\item{object}{An `Model` object.}

\item{max_n}{Integer. The maximum number of rows to print.}

\item{...}{Further arguments passed from other methods}
}
\value{
An object of class `logLik`. If both `fixed` and `covariance` are FALSE then it returns NA.
}
\description{
Summarizes `Model` object.
}
