% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfunctions.R
\name{vcov.mcml}
\alias{vcov.mcml}
\title{Extract the Variance-Covariance matrix for a `mcml` object}
\usage{
\method{vcov}{mcml}(object, ...)
}
\arguments{
\item{object}{A `mcml` object.}

\item{...}{Further arguments passed from other methods}
}
\value{
A variance-covariance matrix.
}
\description{
Returns the calculated variance-covariance matrix for a `mcml` object. The generating Model object 
has several methods to calculate the variance-convaariance matrix. For the standard GLS information matrix see
`Model$information_matrix()`. Small sample corrections 
can be accessed directly from the generating Model using `Model$small_sample_correction()`. The varaince-covariance 
matrix including the random effects can be accessed using `Model$information_matrix(include.re = TRUE)`.
}
