% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_s3_functions.R
\name{vcov.Model}
\alias{vcov.Model}
\title{Calculate Variance-Covariance matrix for a `Model` object}
\usage{
\method{vcov}{Model}(object, ...)
}
\arguments{
\item{object}{A `Model` object.}

\item{...}{Further arguments passed from other methods}
}
\value{
A variance-covariance matrix.
}
\description{
Returns the variance-covariance matrix for a `Model` object. Specifically, this function will 
return the inverse GLS information matrix for the fixed effect parameters. Small sample corrections 
can be accessed directly from the Model using `Model$small_sample_correction()`. The varaince-covariance 
matrix including the random effects can be accessed using `Model$information_matrix(include.re = TRUE)`.
}
