% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{x_names}
\alias{x_names}
\title{Gets the names of the fixed effects variables}
\usage{
x_names(formula)
}
\arguments{
\item{formula}{A string specifying the formula}
}
\value{
A vector of variable names
}
\description{
Gets the names of the fixed effects variables
}
