% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrafunctions.R
\name{blockmat}
\alias{blockmat}
\title{Create block matrix}
\usage{
blockmat(...)
}
\description{
Create a block matrix from the inputs
}
\details{
Takes a sequence of matrices and produces a block matrix.
}
