% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printfunctions.R
\name{fixef.mcml}
\alias{fixef.mcml}
\title{Extracts the fixed effect estimates}
\usage{
\method{ranef}{mcml}(object, ...)
}
\arguments{
\item{object}{An `mcml` model fit.}

\item{...}{Further arguments passed from other methods}
}
\value{
A named, numeric vector of fixed-effects estimates.
}
\description{
Extracts the fixed effect estimates from an mcml object returned from call of `MCML` or `LA` in the \link[glmmrBase]{Model} class.
}
