% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{genCholD}
\alias{genCholD}
\title{Generates the cholesky decomposition covariance matrix of the random effects}
\usage{
genCholD(D_data, gamma)
}
\arguments{
\item{gamma}{Vector of covariance parameters specified in order they appear column wise in the functions 
specified by `func_def`}

\item{B}{Integer specifying the number of blocks in the matrix}

\item{N_dim}{Vector of integers, which each value specifying the dimension of each block}

\item{N_func}{Vector of integers specifying the number of functions in the covariance function 
for each block.}

\item{func_def}{Matrix of integers where each column specifies the function definition for each function in each block.}

\item{N_var_func}{Matrix of integers of same size as `func_def` with each column specying the number 
of variables in the argument to each function in each block}

\item{col_id}{3D array (cube) of integers of dimension length(func_def) x max(N_var_func) x B 
where each slice the respective column indexes of `cov_data` for each function in the block}

\item{N_par}{Matrix of integers of same size as `func_def` with each column specifying the number
of parameters in the function in each block}

\item{sum_N_par}{Total number of parameters}

\item{cov_data}{3D array (cube) holding the data for the covariance matrix where each of the B slices
is the data required for each block}
}
\value{
A lower triangular matrix matrix
}
\description{
Generates the covariance matrix of the random effects from a sparse representation
}
