\name{glmmrBase-package}
\alias{glmmrBase-package}
\alias{glmmrBase}
\docType{package}
\title{
\packageTitle{glmmrBase}
}
\description{
\packageDescription{glmmrBase}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{glmmrBase}
\packageIndices{glmmrBase}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{glmmrBase}

Maintainer: \packageMaintainer{glmmrBase}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
