\name{glmmrBase-package}
\alias{glmmrBase-package}
\alias{glmmrBase}
\docType{package}
\title{
\packageTitle{glmmrBase}
}
\description{
\packageDescription{glmmrBase}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{glmmrBase}
\packageIndices{glmmrBase}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{glmmrBase}

Maintainer: \packageMaintainer{glmmrBase}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
}

