% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_re}
\alias{sample_re}
\title{Generates a sample of random effects}
\usage{
sample_re(formula, data, colnames, theta)
}
\arguments{
\item{formula}{A string specifying the formula}

\item{data}{A matrix with the data}

\item{colnames}{A vector of strings specifying the column names of the data}

\item{theta}{Vector of parameters used to generate the matrix D.}
}
\value{
A symmetric positive definite covariance matrix
}
\description{
Generates a sample of random effects from the specified covariance matrix.
}
